auto_agents_build/requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
numpy>=1.24.0
pyyaml>=6.0
jsonschema>=4.17.0

# Vector store
faiss-cpu>=1.7.4

# Optional: for Chroma vector store
# chromadb>=0.4.0

# Optional: for better performance
# faiss-gpu>=1.7.4
